"importo" parameter for method new Casino(...).incassa(...) is now double instead of int

+++ Java/src/java1/classi/esercizi/giocodadi/Casino.java
@@ -26,7 +26,7 @@ public class Casino {
 		return "Casino " + nome + ", capitale=" + capitale;
 	}
 	
-	public void incassa(int importo) {
+	public void incassa(double importo) {
 		capitale += importo;
 	}
